requirements.txt
matplotlib
pysha3
